+++ src/plugin/KeyPlugin.cpp
@@ -4,12 +4,12 @@ KeyPlugin::KeyPlugin() {
     nCodes = 0;
 }
 
-KeyPlugin::KeyPlugin(int *scanCodes, int nCodes) {
+KeyPlugin::KeyPlugin(uint8_t *scanCodes, int nCodes) {
     this->scanCodes = scanCodes;
     this->nCodes = nCodes;
 }
 
-bool KeyPlugin::isKey(int scanCode) {
+bool KeyPlugin::isKey(uint8_t scanCode) {
     for (int i = 0; i < nCodes; i++) {
         if(scanCodes[i] == scanCode) {
             return true;
